+++ Quicksilver/Code-App/QSHelpersPrefPane.m
@@ -39,7 +39,7 @@
 }
 
 - (void)selectItemInPopUp:(NSPopUpButton *)popUp representedObject:(id)object {
-	int index = [popUp indexOfItemWithRepresentedObject:object];
+	NSUInteger index = [popUp indexOfItemWithRepresentedObject:object];
 	if (index == -1 && [popUp numberOfItems]) index = 0;
 	[popUp selectItemAtIndex:index];
 }
@@ -114,7 +114,7 @@
 	return [helperInfo count];
 }
 
-- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex {
+- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex {
 	if ([[aTableColumn identifier] isEqual:@"helper"]) {
 		return nil;
 	} else {
@@ -122,10 +122,10 @@
 	}
 }
 
-- (void)tableView:(NSTableView *)aTableView setObjectValue:(id)anObject forTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex {
+- (void)tableView:(NSTableView *)aTableView setObjectValue:(id)anObject forTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex {
 	if ([[aTableColumn identifier] isEqual:@"helper"] && anObject) {
 		NSDictionary *info = [helperInfo objectAtIndex:rowIndex];
-		int index = [anObject intValue];
+		NSUInteger index = [anObject intValue];
 		NSMenu *menu = [info objectForKey:MENU];
 		NSDictionary *settings = [info objectForKey:INFO];
 		anObject = [[menu itemAtIndex:index] representedObject];
@@ -139,7 +139,7 @@
 	}
 }
 
-- (void)tableView:(NSTableView *)aTableView willDisplayCell:(NSCell*)aCell forTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex {
+- (void)tableView:(NSTableView *)aTableView willDisplayCell:(NSCell*)aCell forTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex {
 	if ([[aTableColumn identifier] isEqual:@"helper"]) {
 		NSDictionary *info = [helperInfo objectAtIndex:rowIndex];
 		id object = [QSReg getMediatorID:[info objectForKey:IDENT]];
@@ -148,7 +148,7 @@
 		NSMenu *menu = [[helperInfo objectAtIndex:rowIndex] objectForKey:MENU];
 		[aCell setEnabled:[menu numberOfItems] >1];
 		[aCell setMenu:menu];
-		int index = [(NSPopUpButtonCell*)aCell indexOfItemWithRepresentedObject:object];
+		NSUInteger index = [(NSPopUpButtonCell*)aCell indexOfItemWithRepresentedObject:object];
 		if (index == -1 && [(NSPopUpButtonCell*)aCell numberOfItems]) index = 0;
 		[(NSPopUpButtonCell*)aCell selectItemAtIndex:index];
 	}
