+++ Quicksilver/Code-App/QSPlugInsPrefPane.m
@@ -196,7 +196,7 @@
 	NSIndexSet *indexes = [plugInTable selectedRowIndexes];
 	if (!indexes) return nil;
 	NSMutableArray *bundles = [NSMutableArray array];
-	int index;
+	NSInteger index;
 	for (index = [indexes firstIndex]; index != NSNotFound; index = [indexes indexGreaterThanIndex:index]) {
 		[bundles addObject:[[arrayController arrangedObjects] objectAtIndex:index]];
 		if (index == [indexes lastIndex]) break;
@@ -267,7 +267,7 @@
 
 - (IBAction)reloadPlugIns:(id)sender { [[QSPlugInManager sharedInstance] downloadWebPlugInInfoIgnoringDate];  }
 
-- (void)tableView:(NSTableView *)aTableView willDisplayCell:(NSTextFieldCell*)aCell forTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex {
+- (void)tableView:(NSTableView *)aTableView willDisplayCell:(NSTextFieldCell*)aCell forTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex {
 	if ([[aTableColumn identifier] isEqualToString:@"enabled"]) {
 		NSArray *array = [arrayController arrangedObjects];
 		QSPlugIn *object = [array objectAtIndex:rowIndex];
@@ -335,7 +335,7 @@
 
 - (BOOL)isItemExpanded:(id)item {return YES;}
 
-- (float)outlineView:(NSOutlineView *)outlineView heightOfRowByItem:(id)item {
+- (CGFloat)outlineView:(NSOutlineView *)outlineView heightOfRowByItem:(id)item {
 	if ([[[item respondsToSelector:@selector(representedObject)] ? [item representedObject] : [(QSObject *)item observedObject] objectForKey:@"type"] isEqualToString:@"category"])
 		return 16;
 	return 32;
