+++ Quicksilver/Code-QuickStepInterface/QSFileConflictPanel.m
@@ -53,11 +53,11 @@
 	[NSApp stopModal];
 }
 
-- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex {
+- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex {
 	return [[conflictNames objectAtIndex:rowIndex] lastPathComponent];
 }
 
-- (void)tableView:(NSTableView *)aTableView willDisplayCell:(id)aCell forTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex {
+- (void)tableView:(NSTableView *)aTableView willDisplayCell:(id)aCell forTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex {
 	[(QSImageAndTextCell*)aCell setImage:[[NSWorkspace sharedWorkspace] iconForFile:[conflictNames objectAtIndex:rowIndex]]];
 }
